+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Change Log
 
 ## 1.0.9
-- drop support for old connections.alan files
+- drop support for old `connections.alan` files
 - fix context for package command when `deployment.alan` is not opened
 
 ## 1.0.8